requirements.txt
selenium==4.18.1
opencv-python==4.9.0.80
pyzbar==0.1.9
Pillow==10.2.0
python-dotenv==1.0.1
requests==2.31.0
loguru==0.7.2
tqdm==4.66.2
undetected-chromedriver
PyPDF2